Rename rdbstorage to rdb
